refactor(research): validate scenario boundary
