+++ apps/research/src/page.tsx
@@ -1,33 +1,7 @@
-import type { Research } from "@site/data-access";
-import { validateCvDomain } from "@site/data-access";
-import { useEffect, useState } from "react";
 import { ResearchContent } from "./research-content";
+import { useResearch } from "./use-research";
 import "@site/design-system";
 import "./research.css";
-
-type ResearchViewState =
-  | { name: "loading" }
-  | { name: "error" }
-  | { name: "ready"; research: Research };
-
-function researchUrl() {
-  const url = new URL(
-    "/nick-derobertis-site/cv-data/domains/research.json",
-    window.location.origin,
-  );
-  const scenario = new URLSearchParams(window.location.search).get(
-    "research-scenario",
-  );
-  if (scenario) url.searchParams.set("scenario", scenario);
-  return url;
-}
-
-async function loadResearch(signal: AbortSignal) {
-  const response = await fetch(researchUrl(), { signal });
-  if (!response.ok)
-    throw new Error(`Research request failed: ${response.status}`);
-  return validateCvDomain("research", await response.json());
-}
 
 function StateMessage({ state }: { state: "empty" | "loading" | "error" }) {
   const messages = {
@@ -48,18 +22,7 @@ function StateMessage({ state }: { state: "empty" | "loading" | "error" }) {
 }
 
 export default function ResearchPage() {
-  const [state, setState] = useState<ResearchViewState>({ name: "loading" });
-  useEffect(() => {
-    const controller = new AbortController();
-    loadResearch(controller.signal).then(
-      (research) => setState({ name: "ready", research }),
-      (error: unknown) => {
-        if (!(error instanceof DOMException && error.name === "AbortError"))
-          setState({ name: "error" });
-      },
-    );
-    return () => controller.abort();
-  }, []);
+  const state = useResearch();
   if (state.name !== "ready") return <StateMessage state={state.name} />;
   if (!state.research.projects?.length) return <StateMessage state="empty" />;
   return <ResearchContent research={state.research} />;
